Created basic walking and item getting simulator
The main file is where all of the game input happens, but the
src package is where all the actual commands are. I have
test_world set up to do very basic testing for the commands.

In the future I'm going to be cleaning up the item methods, as
they are kind of messy. After that I'll be adding more item
and travel methods, as well as creating methods for puzzles
and eventually dialogue. See todolist.txt.