Basic wearing system implemented. New test world.
I added some clothing template classes for ease of item creation.
In doing so I had to change around a lot of the functions for the
player, rooms, and containers since they threw errors.

Additionally I added the ability to open and close inventories in
rooms. This also required some careful refactoring of the find and
search item codes in order to fit the closed and open inventories.

I noticed that when I had more than one inventory in a room, like
the starting cabin, I couldn't do 'get item' from the chest, and
it only worked if I specified the chest. Turns out that the find
item function was resetting the found item every time it looped
through a new sub-container. So, I had to fix that.