Added a Backpack to the test world.
Backpacks were instantiated just like a regular item, but the
language around backpacks required me to add all containers as
items to an inventory. This created some problems for room
descriptions that I will work on in a later patch. Really want
to get items worked out first before I move on to the next thing.