Added Wearable Containers
This modification was a big one. I had to add an attribute to all
items to check if it was a container or not. Then I cleaned up the
find_item method to be a lot more readable and more efficient,
until I pour more spaghetti onto that mess. That's a problem for
the next patch, though.

What I did for find item was add all the containers up into an
array, and then iterated over every container, instead of having
to recursively run find_item over every single container. That's
a big reason why the item_id = False kept breaking the method.

Finally, I added make_table under the person class, so that I can
display multiple bits of info about the PC or NPC on the same screen
instead of creating a scroll.