Added a Bag as an Item Class.
A bag is basically a container that can be worn. Eventually this
will be expanded to pouches. I'm thinking that instead of making
holsters a type of bag, I will make them their own separate type
of Wearable that only accepts one type of item. For example, a
sheathe can only accept a sword. I can expand that "holster" code
to a weapon rack.